Added viewControllers for all the screens, CustomGradient extension is added.
