+++ SnipApp25/ForgotPasswordViewController.swift
@@ -1,14 +1,15 @@
 //
-//  LoginViewController.swift
+//  ForgotPasswordViewController.swift
 //  SnipApp25
 //
-//  Created by Srikanth Kyatham on 11/11/24.
+//  Created by Srikanth Kyatham on 11/12/24.
 //
 
 import Foundation
+import Foundation
 import UIKit
 
-class LoginViewController : UIViewController {
+class ForgotPasswordViewController : UIViewController {
     
     override func viewDidLoad() {
         super.viewDidLoad()
